Add files via upload
Several minor, mostly inconsequential at this point, changes to the classification and grid search.
Mainly, created a list of the KMeans cluster fits for different numbers of clusters. 
This is with a view to using cluster membership to create a balanced training set for the classification algorithms in the future.