Add files via upload
chore: Changed read_saved_categorization flag to False

Set the flag in preparation for final project submission. Setting to False will run the all the code in the notebook since it will be the first time the reviewer runs the code. This means the code will take longer to execute. Setting to True skips some time-consuming computations and reads stored intermediate results.